+++ main.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"regexp"
 	"strings"
 	"text/template"
 
@@ -15,6 +16,10 @@ import (
 	"github.com/johndistasio/cauldron/version"
 
 	"github.com/Masterminds/sprig"
+)
+
+var (
+	httpRegex = regexp.MustCompile("^http(s){0,1}://")
 )
 
 func init() {
@@ -33,8 +38,8 @@ Arguments:
         Render the template in-place (overwriting the template) instead of to
         standard output. Takes precedence over -file.
     -json:
-        Read template data from the specified JSON file. Command-line template
-        parameters are ignored.
+        Read template data from the specified JSON file or URL. Command-line
+		template parameters are ignored.
     -template:
         Path to the template to be rendered. This argument is required.
     -version:
@@ -55,8 +60,8 @@ Example:
 	}
 }
 
-func renderTemplate(t *template.Template, p data.Provider) ([]byte, error) {
-	d, err := p.GetData()
+func renderTemplate(t *template.Template, s data.Source) ([]byte, error) {
+	d, err := s.GetData()
 
 	if err != nil {
 		err = errors.New(fmt.Sprintf("failed to parse data: %s", err.Error()))
@@ -114,16 +119,23 @@ func main() {
 		quit(err)
 	}
 
-	var prv data.Provider
-	var file *os.File
-	defer file.Close()
+	var src data.Source
 
 	switch {
+	case httpRegex.MatchString(*jsonPtr):
+		src = data.NewHttp(*jsonPtr, nil)
 	case *jsonPtr != "":
-		prv = data.NewJsonFile(*jsonPtr)
+		src = data.NewJsonFile(*jsonPtr)
 	default:
-		prv = data.NewCommandLine(flag.Args())
+		src = data.NewCommandLine(flag.Args())
 	}
+
+	if err != nil {
+		quit(err)
+	}
+
+	var file *os.File
+	defer file.Close()
 
 	switch {
 	case *inplacePtr:
@@ -138,7 +150,7 @@ func main() {
 		quit(err)
 	}
 
-	rendered, err := renderTemplate(tmpl, prv)
+	rendered, err := renderTemplate(tmpl, src)
 
 	if err != nil {
 		quit(err)